Bugfix: Vergleich von Long-Werten in patch-Methoden korrigiert
closes #674

+++ svws-openapi/src/main/java/de/nrw/schule/svws/data/betriebe/DataBetriebAnsprechpartner.java
@@ -86,33 +86,28 @@ public class DataBetriebAnsprechpartner  extends DataManager<Long>{
 			try {
 				conn.transactionBegin();
 				DTOAnsprechpartnerAllgemeineAdresse ansprechpartner = conn.queryByKey(DTOAnsprechpartnerAllgemeineAdresse.class, id);
-				
 				if ( ansprechpartner == null)
 					throw OperationError.NOT_FOUND.exception();
-				
 				for(Entry<String, Object> entry : map.entrySet()) {
 					String key = entry.getKey();
 					Object value = entry.getValue();
 					switch(key) {
-						
 						case "id" -> {
 							Long patch_id = JSONMapper.convertToLong(value, true) ;
-							if( (patch_id == null) || (patch_id.intValue()  != id))
+							if( (patch_id == null) || (patch_id.intValue() != id.intValue()))
 								throw OperationError.BAD_REQUEST.exception();
 						}
-						
 						case "betrieb_id" -> {
 							Long betrieb_id = JSONMapper.convertToLong( value , true );
-							if(betrieb_id == null) {
+							if (betrieb_id == null) {
 								ansprechpartner.Adresse_ID = null;
-							}else {
+							} else {
 								DTOKatalogAllgemeineAdresse betrieb = conn.queryByKey(DTOKatalogAllgemeineAdresse.class, betrieb_id);
 								if( betrieb == null)
 									throw OperationError.NOT_FOUND.exception();
 								ansprechpartner.Adresse_ID = betrieb.ID;
 							}
 						}
-						
 						case "titel" -> ansprechpartner.Titel = JSONMapper.convertToString(value, true, true);
 						case "anrede" -> ansprechpartner.Anrede = JSONMapper.convertToString(value, true, true);
 						case "name" -> ansprechpartner.Name = JSONMapper.convertToString(value, true, true);
@@ -121,18 +116,17 @@ public class DataBetriebAnsprechpartner  extends DataManager<Long>{
 						case "telefon" -> ansprechpartner.Telefon = JSONMapper.convertToString(value, true, true);
 						case "abteilung" -> ansprechpartner.Abteilung= JSONMapper.convertToString(value, true, true);
 						case "GU_ID" -> ansprechpartner.GU_ID= JSONMapper.convertToString(value, true, true);
-					
 						default -> throw OperationError.BAD_REQUEST.exception();
 					}
 				}
 				conn.transactionPersist(ansprechpartner);
-				if( !conn.transactionCommit())
+				if (!conn.transactionCommit())
 					return OperationError.CONFLICT.getResponse("Datenbankfehler beim Persistieren des Betriebansprechpartners");
-			}catch (Exception e) {
+			} catch (Exception e) {
 				if(e instanceof WebApplicationException webApplicationException)
 					return webApplicationException.getResponse();
 				return OperationError.INTERNAL_SERVER_ERROR.getResponse();
-			}finally {
+			} finally {
 				conn.transactionRollback();
 			}
 		}
@@ -157,7 +151,7 @@ public class DataBetriebAnsprechpartner  extends DataManager<Long>{
 					DTODBAutoInkremente lastID = conn.queryByKey(DTODBAutoInkremente.class, "AllgAdrAnsprechpartner");
 					Long ID = lastID == null ? 1 : lastID.MaxID + 1;
 					
-					if(betrieb_id != null) {
+					if (betrieb_id != null) {
 						DTOKatalogAllgemeineAdresse betrieb = conn.queryByKey(DTOKatalogAllgemeineAdresse.class, betrieb_id);
 						if( betrieb == null)
 							throw OperationError.NOT_FOUND.exception("Ein Betrieb mit der ID "+betrieb_id+" existiert in der Datenbank nicht.");
@@ -165,21 +159,20 @@ public class DataBetriebAnsprechpartner  extends DataManager<Long>{
 					// Ansprechpartner anlegen
 					ansprechpartner = new DTOAnsprechpartnerAllgemeineAdresse(ID, betrieb_id);
 					
-					for(Entry<String, Object> entry : map.entrySet()) {
+					for (Entry<String, Object> entry : map.entrySet()) {
 						String key = entry.getKey();
 						Object value = entry.getValue();
 						switch(key) {
-							
 							case "id" -> {
 								Long create_id = JSONMapper.convertToLong(value, true) ;
-								if ( create_id != null )
+								if (create_id != null)
 									throw OperationError.BAD_REQUEST.exception("Databetrieb_Ansprechpartner-ID muss bei der Erstellung null sein.");
 							}
 							case "betrieb_id" -> {
 								Long bid = JSONMapper.convertToLong( value , true );
 								if (bid == null)
 									throw OperationError.BAD_REQUEST.exception("SchülerID darf nicht fehlen.");
-								if (bid != betrieb_id)
+								if (bid.longValue() != betrieb_id.longValue())
 									throw OperationError.BAD_REQUEST.exception("Betrieb_ID aus dem JSON-Objekt stimmt mit dem übergebenen Argument nicht überein.");
 							}
 							case "titel" -> ansprechpartner.Titel = JSONMapper.convertToString(value, true, true);
