+++ svws-openapi/src/main/java/de/nrw/schule/svws/data/betriebe/DataKatalogBeschaeftigunsarten.java
@@ -100,10 +100,10 @@ public class DataKatalogBeschaeftigunsarten extends DataManager<Long> {
                 for (Entry<String, Object> entry : map.entrySet()){
                     String key = entry.getKey();
                     Object value = entry.getValue();
-                    switch (key){
+                    switch (key) {
 						case "id" -> {
 							Long patch_id = JSONMapper.convertToLong(value, true);
-							if ((patch_id == null) || (patch_id != id))
+							if ((patch_id == null) || (patch_id.longValue() != id.longValue()))
 								throw OperationError.BAD_REQUEST.exception();
 						}
                         case "text" -> art.Bezeichnung = JSONMapper.convertToString(value, true, true);
