+++ svws-openapi/src/main/java/de/nrw/schule/svws/data/erzieher/DataErzieherStammdaten.java
@@ -152,7 +152,7 @@ public class DataErzieherStammdaten extends DataManager<Long> {
                     switch (key){
 						case "id" -> {
 							Long patch_id = JSONMapper.convertToLong(value, true);
-							if ((patch_id == null) || (patch_id != tmpid))
+							if ((patch_id == null) || (patch_id.longValue() != tmpid.longValue()))
 								throw OperationError.BAD_REQUEST.exception();
 						}
                         case "idSchueler" -> throw OperationError.BAD_REQUEST.exception();
