+++ svws-openapi/src/main/java/de/nrw/schule/svws/data/lehrer/DataLehrerStammdaten.java
@@ -108,7 +108,7 @@ public class DataLehrerStammdaten extends DataManager<Long> {
 		    			// Basisdaten
 						case "id" -> {
 							Long patch_id = JSONMapper.convertToLong(value, true);
-							if ((patch_id == null) || (patch_id != id))
+							if ((patch_id == null) || (patch_id.longValue() != id.longValue()))
 								throw OperationError.BAD_REQUEST.exception();
 						}
 		    			case "foto" -> {
