+++ svws-openapi/src/main/java/de/nrw/schule/svws/data/schueler/DataSchuelerSchulbesuchsdaten.java
@@ -133,7 +133,7 @@ public class DataSchuelerSchulbesuchsdaten extends DataManager<Long> {
 		    		switch (key) {
 						case "id" -> {
 							Long patch_id = JSONMapper.convertToLong(value, true);
-							if ((patch_id == null) || (patch_id != id))
+							if ((patch_id == null) || (patch_id.longValue() != id.longValue()))
 								throw OperationError.BAD_REQUEST.exception();
 						}
